requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.5
colorama==0.4.6
distlib==0.3.7
filelock==3.12.2
flash==1.0.3
Flask==2.3.2
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
passlib==1.7.4
platformdirs==3.10.0
protobuf==4.21.12
PyMySQL==1.1.0
pytz==2023.3
six==1.16.0
virtualenv==20.24.3
Werkzeug==2.3.6
